+++ HISTORY.md
@@ -11,6 +11,9 @@
   * Use log-space to represent HMM initial state and transition probabilities
     (#2081).
 
+  * Add functions to access parameters of `Convolution` and `AtrousConvolution`
+    layers (#1985).
+
 ### mlpack 3.2.2
 ###### 2019-11-26
   * Add `valid` and `same` padding option in `Convolution` and `Atrous